requirements.txt

# Core ML/DL frameworks
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0

# Scientific computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Image processing
Pillow>=8.3.0
pydicom>=2.3.0

# Visualization and monitoring
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.7.0
tqdm>=4.62.0

# Utilities
joblib>=1.1.0
logging>=0.4.9.6

# Optional: for better performance
# torch-audio>=0.9.0  # if audio processing needed
# opencv-python>=4.5.0  # alternative image processing